Migrate CmdPreprocessor tests to testsuite
